feat: 통합 유효성검사 모듈을 구현한다
