+++ packages/generate-design-token/src/index.ts
@@ -4,9 +4,8 @@ export { default as generateDesignToken } from "./generate-design-token";
 export * from "./types/token.types";
 export * from "./modules/transform-use-cases";
 export * from "./modules/transform";
+export * from "./modules/validate";
 export * from "./utils/token-obj";
 export * from "./utils/token-ref";
 export * from "./utils/validate/format";
-export * from "./utils/validate/token/duplicate";
-export * from "./utils/validate/token/is";
-export * from "./utils/validate/token/validate";
+export * from "./utils/validate/duplicate";
